refactor(helpers): organize helpers in dir
