+++ lib/bitcoin/block.rb
@@ -1,6 +1,6 @@
 require_relative '../bitcoin_data_io'
-require_relative '../encoding_helper'
-require_relative '../hash_helper'
+require_relative '../helpers/encoding_helper'
+require_relative '../helpers/hash_helper'
 
 module Bitcoin
   class Block
