+++ lib/helpers/encoding_helper.rb
