+++ lib/helpers/hash_helper.rb
